% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_ingredients.R
\name{activeingredient_api}
\alias{activeingredient_api}
\title{Query Active Ingredient API of the Drug Product Database by}
\usage{
activeingredient_api(active_ingredient = NULL, drug_code = NULL)
}
\arguments{
\item{active_ingredient}{Active ingredient (string or vector). Enter term just as you would in the online DPD query (not case-sensitive). Only supply one term.}

\item{drug_code}{Code of the drug product.}
}
\value{
Response from API call as a list.
}
\description{
A function that queries the Active Ingredient API of the DPD using drug codes or search terms of active ingredients.
Use \code{\link[=search_ai_api]{search_ai_api()}} if you wish to query the API using a vector of multiple values.
}
\examples{
# activeingredient_api() # returns all values in API
# activeingredient_api('semaglutide', drug_code=NULL)
# library(purrr)
# map(c('semaglutide', 'metformin'), activeingredient_api)
}
\seealso{
\code{\link[=search_ai_api]{search_ai_api()}}
}
\author{
Ezinne MM Ndukwe
}
