% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosage_form.R
\name{dosageform_api}
\alias{dosageform_api}
\title{Query Dosage Form API of the Drug Product Database}
\usage{
dosageform_api(drug_code, active_param = NULL)
}
\arguments{
\item{drug_code}{Code of the drug product}

\item{active_param}{Dosage form(s) that are active. Default is NULL. Use 'yes' to return the dosage form(s) that have a date that is greater than today or no date.}
}
\value{
Response from API call as a list.
}
\description{
This function queries the Dosage Form API using drug codes
Use \code{\link[=search_dosageform_api]{search_dosageform_api()}} if you wish to query the API using multiple drug codes.
}
\examples{
dosageform_api(drug_code = 3847)
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}
}
\author{
Ezinne MM Ndukwe
}
