% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_products.R
\name{drugproduct_api}
\alias{drugproduct_api}
\title{Query Drug Product API of the Drug Product Database}
\usage{
drugproduct_api(
  drug_code = NULL,
  drug_identification_number = NULL,
  brand_name = NULL,
  status = NULL
)
}
\arguments{
\item{drug_code}{Code of the drug product.}

\item{drug_identification_number}{Drug Identification Number (DIN) of the drug product as a string.}

\item{brand_name}{Brand name of the drug product.}

\item{status}{Drug product status.}
}
\value{
Response from API call as a list
}
\description{
A function that queries Drug Product API using drug code, DIN, brand name or status.
Results can be unpredictable if more than one search parameters (drug code, DIN and brand name) are used. Recommended to use only one.
}
\examples{
drugproduct_api() # all drug products in the DPD
drugproduct_api(drug_code = 3847)
drugproduct_api(drug_identification_number = '00446564')
}
\seealso{
\code{\link[=search_dp_api]{search_dp_api()}}
}
\author{
Ezinne MM Ndukwe
}
