% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_and_bind_list}
\alias{name_and_bind_list}
\title{Name and bind list}
\usage{
name_and_bind_list(list, names = NULL, names_to = NULL)
}
\arguments{
\item{list}{List (e.g., results of a mapped API query)}

\item{names}{Vector used to name the elements of a list or atomic vector}

\item{names_to}{Name of an optional identifier column. If value is supplied to names parameter, this parameter defaults to 'name'.}
}
\value{
Bound and named list
}
\description{
The name_and_bind function uses the names() and dplyr:bind_rows() functions to assigns names to lists (if names parameter is used) and bind list elements into a dataframe.
}
\details{
Name and bind
}
\examples{
#library(purrr)
#list_of_drugs <- c('semaglutide', 'insulin')
#api_result <- purrr::map(list_of_drugs, ~activeingredient_api(active_ingredient = .x))
#name_and_bind_list(api_result)
#name_and_bind_list(api_result, names = list_of_drugs, names_to = 'drug_class')
#name_and_bind_list(api_result, names = list_of_drugs) # Default name of identifier column is 'name'
#name_and_bind_list(api_result, names_to = 'drug_class')
}
