% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{output_valid_results}
\alias{output_valid_results}
\title{Output (retain) valid responses of an API response}
\usage{
output_valid_results(
  named_list,
  second_element_to_check = NULL,
  name_source = NULL,
  api_name = NULL,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{named_list}{The named list generated from an API call.}

\item{second_element_to_check}{Second-level list element name (or nested list element name) as a string to check. If the value is null then element name is considered not found}

\item{name_source}{Optional string to say what the names represent (e.g., "drug codes").}

\item{api_name}{Optional. Name of API. Used for alert if alert_val_not_found is TRUE.}

\item{alert_val_not_found}{Show alert if a value of the supplied search parameter is not in the list. If TRUE, alerts show (default). Use FALSE to hide.}
}
\description{
This function evaluates the responses of an API call and retains the valid responses in the output.
}
\examples{
#library(purrr)
#drug_codes_list <- c(0, 1, 8, 35413)
#api_results <- map(drug_codes_list, ~drugproduct_api(drug_code = .x))
#names(api_results) <- drug_codes_list
#output_valid_results(named_list = api_results, second_element_to_check = 'drug_identification_number', name_source = 'drug codes', api_name = 'Drug Products')
}
