% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_status.R
\name{productstatus_api}
\alias{productstatus_api}
\title{Query Product Status API of the Drug Product Database}
\usage{
productstatus_api(drug_code = NULL)
}
\arguments{
\item{drug_code}{Code of the drug product}
}
\value{
Results of the API query as a list.
}
\description{
A function that queries Product Statuses API using drug codes.
}
\details{
Use \code{\link[=search_productstatus_api]{search_productstatus_api()}} if you wish to query the API using a vector of multiple drug codes.
}
\examples{
productstatus_api(drug_code = 3847)
}
\author{
Ezinne MM Ndukwe
}
