% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dpd_by_atc.R
\name{query_by_atc}
\alias{query_by_atc}
\title{Query DPD by Anatomical Therapeutic Chemical (ATC) code}
\usage{
query_by_atc(
  atc,
  active_param = NULL,
  nest_additional_info = FALSE,
  progress_bar = TRUE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{atc}{Anatomical Therapeutic Chemical (ATC) code (as string or vector of strings)}

\item{active_param}{Only return dosage forms, schedules and routes of administration that are active. Default is NULL. Use 'yes' to returns parameter values that have a date that is greater than today or no date.}

\item{nest_additional_info}{If TRUE, some columns are hidden (nested) under column 'additional_info'. This is similar to how you need to click on the hyperlinked DIN to show more information for a drug product in the search results summary of the DPD online query.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a drug code is not in a component API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\description{
A function that queries the entire DPD by ATC and returns a data frame of the results.
}
\examples{
query_by_atc(atc = c('R03DC03', 'B01AC27', '0'), nest_additional_info = TRUE)
query_by_atc(atc = c('R03DC03 ', '0'), nest_additional_info = TRUE) # adding space to 'R03DC03' to cause invalid value
}
