% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dpd_by_brandName.R
\name{query_dpd_by_brandName}
\alias{query_dpd_by_brandName}
\title{Query DPD by brand name}
\usage{
query_dpd_by_brandName(
  brand_name,
  active_param = NULL,
  nest_additional_info = FALSE,
  progress_bar = TRUE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{brand_name}{Brand name (or product name) of the drug product. Can be string or vector of strings.}

\item{nest_additional_info}{If TRUE, some columns are hidden (nested) under column 'additional_info'. This is similar to how you need to click on the hyperlinked DIN to show more information for a drug product in the search results summary of the DPD online query.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a drug code is not in a component API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\description{
A function that queries the DPD by brand names(s) and returns a data frame of the results.
}
\examples{
query_dpd_by_brandName(c('ozempic', 'metamucil'), nest_additional_info = TRUE)
}
\author{
Ezinne MM Ndukwe
}
