% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dpd_by_din.R
\name{query_dpd_by_din}
\alias{query_dpd_by_din}
\title{Query DPD by Drug Identification Number (DIN)}
\usage{
query_dpd_by_din(
  drug_identification_number,
  active_param = NULL,
  nest_additional_info = FALSE,
  progress_bar = TRUE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_identification_number}{DIN of the drug product. Can be one DIN as string or vector of strings.}

\item{active_param}{Only return dosage forms, schedules and routes of administration that are active. Default is NULL. Use 'yes' to returns parameter values that have a date that is greater than today or no date.}

\item{nest_additional_info}{If TRUE, some columns are hidden (nested) under column 'additional_info'. This is similar to how you need to click on the hyperlinked DIN to show more information for a drug product in the search results summary of the DPD online query.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a DIN or drug code is not in a component API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\description{
A function that queries the DPD by DIN(s) and returns a data frame of the results.
}
\examples{
query_dpd_by_din(drug_identification_number = '02416794')
query_dpd_by_din(drug_identification_number = c('0', '02416794'))
query_dpd_by_din(drug_identification_number = c('0')) # should come up blank
query_dpd_by_din(
  drug_identification_number = c('00446580', '53756', '90082', '89002'),
  active_param='yes'
  ) # Retrieve active dosage forms, schedules and routes of administration; these products are all canceled now
}
\author{
Ezinne MM Ndukwe
}
