% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dpd_by_drugCode.R
\name{query_dpd_by_drugCode}
\alias{query_dpd_by_drugCode}
\title{Query Drug Product Database by drug code}
\usage{
query_dpd_by_drugCode(
  drug_code,
  active_param = NULL,
  nest_additional_info = FALSE,
  progress_bar = TRUE,
  alert_val_not_found = TRUE,
  include_tc = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{active_param}{Only return dosage forms, schedules and routes of administration that are active. Default is NULL. Use 'yes' to returns parameter values that have a date that is greater than today or no date.}

\item{nest_additional_info}{If TRUE, additional columns are hidden (nested) under column 'additional_info'.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default). Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a drug code is not in a component API. If TRUE, alerts show (default). Use FALSE to hide.}

\item{include_tc}{Include therapeutic class information (e.g., Anatomical Therapeutic Chemical Code).}
}
\description{
A function that queries the DPD by drug codes(s) and returns a data frame of the results.
}
\examples{
# query_dpd_by_drugCode(96058)
# query_dpd_by_drugCode(c(53756), active_param='yes') # Cancelled post-market
}
