% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_of_administration.R
\name{routeadminstration_api}
\alias{routeadminstration_api}
\title{Query Route of Administration API of the Drug Product Database}
\usage{
routeadminstration_api(drug_code, active_param = NULL)
}
\arguments{
\item{drug_code}{Code of the drug product}

\item{active_param}{Routes of administration that are active. Default is NULL. Use 'yes' to Returns the routes of administration that have a date that is greater than today or no date.}
}
\value{
Results of the API query using the supplied parameter(s).
}
\description{
routeadminstration_api() queries the Route of Administration API using drug codes
}
\details{
Search Route of Administration API
}
\examples{
routeadminstration_api(drug_code = 3847)
}
\author{
Ezinne MM Ndukwe
}
