% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule_api}
\alias{schedule_api}
\title{Query Schedule API of the Drug Product Database}
\usage{
schedule_api(drug_code = NULL, active_param = NULL)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{active_param}{Default is NULL. Use 'yes' to returns the schedule that has a date that is greater than today or no date.}
}
\value{
Results of the API query using the supplied parameter(s).
}
\description{
A function that queries Schedule API using drug codes
}
\details{
Use \code{\link[=search_schedule_api]{search_schedule_api()}} if you wish to query the API using a vector of multiple drug codes.
}
\examples{
schedule_api(drug_code = 3847)
search_schedule_api(drug_code = list(8, 9))
schedule_api(drug_code = 3847, active_param='yes')
}
\author{
Ezinne MM Ndukwe
}
