% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_ingredients.R
\name{search_ai_api}
\alias{search_ai_api}
\title{Return data frame of results from Active Ingredients API query of the Drug Product Database}
\usage{
search_ai_api(
  active_ingredient = NULL,
  drug_code = NULL,
  return_search_terms = TRUE,
  summarize_drugs = TRUE,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{active_ingredient}{Active ingredient(s) to search in the API. Can be string or vector of strings.}

\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{return_search_terms}{Retain column of the supplied active ingredients terms in the output data frame.}

\item{summarize_drugs}{Summarize all found active ingredients for the drug product in one row. Search the Active Ingredients API by drug code to ensure all active ingredients in the drug product are found.}

\item{progress_bar}{Show progress bar. If TRUE (default), progress bar shows with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a value of the supplied search parameter is not in the API. If TRUE (default), alerts show. Use FALSE to hide.}
}
\value{
Data frame of the results. Columns are formatted to match the output of the Drug Product Database online query.
}
\description{
A function that queries the Active Ingredients API and outputs a data frame. Unlike \code{\link[=activeingredient_api]{activeingredient_api()}}, vectors can be supplied to drug codes or active ingredients parameters to query the API with multiple values.
One of drug codes or active ingredients need to be non-null. The two parameters should not be used in combination as the API behaves unpredictably.
Run \code{\link[=activeingredient_api]{activeingredient_api()}} if you wish to return all possible entries in the API.
}
\examples{
search_ai_api(c('metformin', 'lixisenatide', 'sitagliptin'))
search_ai_api(drug_code = 96906)
search_ai_api(drug_code = 96906, summarize_drugs=FALSE)
}
\seealso{
\code{\link[=activeingredient_api]{activeingredient_api()}}, , \code{\link[=query_dpd_by_activeIngredient]{query_dpd_by_activeIngredient()}}
}
