% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosage_form.R
\name{search_dosageform_api}
\alias{search_dosageform_api}
\title{Return data frame of results from Dosage Form API of the Drug Product Database}
\usage{
search_dosageform_api(
  drug_code,
  active_param = NULL,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{active_param}{Dosage form(s) that are active. Default is NULL. Use 'yes' to return the dosage form(s) that have a date that is greater than today or no date.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a value of the supplied search parameter is not in the API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\value{
Data frame of the results. Columns are formatted to match the output of the Drug Product Database online query
}
\description{
A function that queries the Dosage Form API and outputs a data frame. Unlike \code{\link[=dosageform_api]{dosageform_api()}}, vectors can be supplied to drug codes parameter to query the API with multiple values.
}
\examples{
search_dosageform_api(drug_code = 96906)
search_dosageform_api(drug_code = 96906, active_param = 'yes')
}
\seealso{
\code{\link[=dosageform_api]{dosageform_api()}}
}
