% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_products.R
\name{search_dp_api}
\alias{search_dp_api}
\title{Return data frame of results from Drug Product API of the Drug Product Database}
\usage{
search_dp_api(
  drug_code = NULL,
  drug_identification_number = NULL,
  brand_name = NULL,
  find_ais = TRUE,
  progress_bar = TRUE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{drug_identification_number}{Drug Identification Number (DIN) of the drug product. Can be string or vector of strings.}

\item{brand_name}{Brand name (or product name) of the drug product. Can be string or vector of strings.}

\item{find_ais}{Search DPD to find all active ingredients of the identified drug product. Default is TRUE to find.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a value of the supplied search parameter is not in the API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\value{
Dataframe of the results. Columns are formatted to match the output of the Drug Product Database online query
}
\description{
A function that queries Drug Product API using drug codes, DINs or brand names and returns a data frame of the results.
Unlike \code{\link[=drugproduct_api]{drugproduct_api()}}, vectors can be supplied to search parameters to query the API with multiple values.
}
\details{
Results can be unpredictable if more than one search parameters (drug code, DIN and brand name) are used. Only one use one search parameter; the rest should be NULL.
}
\examples{
search_dp_api(drug_code = 3847)
search_dp_api(drug_code = c("90140", "3847"))
search_dp_api(drug_identification_number = c("00446564"))
search_dp_api(brand_name = c("Ozempic"))
search_dp_api(brand_name = "tobradex") # not case-sensitive

# Get drug product info for all drug products in DPD
# search_dp_api()
}
\seealso{
\code{\link[=drugproduct_api]{drugproduct_api()}}
}
\author{
Ezinne MM Ndukwe
}
