% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_status.R
\name{search_productstatus_api}
\alias{search_productstatus_api}
\title{Return data frame of results from Product Status API of the Drug Product Database}
\usage{
search_productstatus_api(
  drug_code,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values. Cannot bu NULL}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if any of the supplied drug code(s) is not in the API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\value{
Dataframe of the results. Columns are formatted to match the output of the Drug Product Database online query.
}
\description{
A function that queries the Product Status API using drug codes and returns a data frame of the results. Unlike productstatus_api(), vectors can be supplied to drug codes parameter to query the API with multiple values.
}
\examples{
search_productstatus_api(drug_code = 3847)
}
\seealso{
\code{\link[=productstatus_api]{productstatus_api()}}
}
\author{
Ezinne MM Ndukwe
}
