% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_of_administration.R
\name{search_routeadminstration_api}
\alias{search_routeadminstration_api}
\title{Return data frame of results from TRoute of Administration API of the Drug Product Database}
\usage{
search_routeadminstration_api(
  drug_code,
  active_param = NULL,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product.}

\item{active_param}{Routes of administration that are active. Default is NULL. Use 'yes' to Returns the routes of administration that have a date that is greater than today or no date.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if any of the supplied drug code(s) is not in the API. If TRUE, alerts show. Use FALSE to hide.}
}
\value{
Data frame of the results. Columns are formatted to match the output of the Drug Product Database online query.
}
\description{
Return data frame of results from TRoute of Administration API of the Drug Product Database
}
\examples{
# search_routeadminstration_api(8)


}
