% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{search_schedule_api}
\alias{search_schedule_api}
\title{Return data frame of results from Schedule API query}
\usage{
search_schedule_api(
  drug_code,
  active_param = NULL,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Code of the drug product. Can be string, numeric, vector or list of values}

\item{active_param}{Default is NULL. Use 'yes' to returns the schedule that has a date that is greater than today or no date.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default) with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if any of the supplied drug code(s) is not in the API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\value{
Data frame of the results. Columns are formatted to match the output of the Drug Product Database online query.
}
\description{
A function that queries the Schedule API using drug codes and returns a data frame of the results. Unlike schedule_api(), vectors can be supplied to the drug codes parameter to query the API with multiple values.
}
\details{
drug_code is a required parameter for this function. Use \code{\link[=schedule_api]{schedule_api()}} if you wish to return all possible drug codes in the API.
}
\examples{
search_schedule_api(drug_code = 3847)
}
\seealso{
\code{\link[=schedule_api]{schedule_api()}}
}
\author{
Ezinne MM Ndukwe
}
