% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therapeutic_class.R
\name{search_therapeuticclass_api}
\alias{search_therapeuticclass_api}
\title{Return data frame of results from Therapeutic Class API of the Drug Product Database}
\source{
\url{https://health-products.canada.ca/api/documentation/dpd-documentation-en.html}
}
\usage{
search_therapeuticclass_api(
  drug_code,
  progress_bar = FALSE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Required. Code of the drug product. Can be string, numeric, vector or list of values. Cannot be NULL.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows with progress messages. Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if any of the supplied drug code(s) is not in the API. If TRUE, alerts show. Use FALSE to hide.}
}
\value{
Data frame of the results. Columns are formatted to match the output of the Drug Product Database online query
}
\description{
A function that queries the Therapeutic Class API using drug codes and returns a data frame of the results. Unlike \code{\link[=therapeuticclass_api]{therapeuticclass_api()}}, vectors can be supplied to drug codes parameter to query the API with multiple values.
}
\examples{
search_therapeuticclass_api(drug_code = 3847)
}
\seealso{
\code{\link[=therapeuticclass_api]{therapeuticclass_api()}}, \code{\link[=query_dpd_by_atc]{query_dpd_by_atc()}}
}
\author{
Ezinne MM Ndukwe
}
