% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_mapping.R
\docType{data}
\name{status_mapping}
\alias{status_mapping}
\title{Mapping table for status codes and descriptions}
\format{
\code{status_mapping}: a data frame with 15 rows and 2 columns:
\describe{
\item{status_code}{Status code}
\item{status_name}{Status name}
}
}
\source{
\url{https://health-products.canada.ca/api/documentation/dpd-documentation-en.html#a4.1}
}
\usage{
status_mapping
}
\description{
A mapping table for the status codes and descriptions in the Drug Product Database.
}
\keyword{datasets}
