% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therapeutic_class.R
\name{therapeuticclass_api}
\alias{therapeuticclass_api}
\title{Query Therapeutic Class API of the Drug Product Database}
\usage{
therapeuticclass_api(drug_code = NULL)
}
\arguments{
\item{drug_code}{Code of the drug product.}
}
\value{
Results of the API query using the supplied parameter(s).
}
\description{
A function that queries Therapeutic Class API using drug codes.
}
\details{
Use \code{\link[=search_therapeuticclass_api]{search_therapeuticclass_api()}} if you wish to query the API using a vector of multiple drug codes.
}
\examples{
therapeuticclass_api(drug_code = 3847)
}
\author{
Ezinne MM Ndukwe
}
