% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_drugCodes}
\alias{use_drugCodes}
\title{Query Product Status, Schedule, Route of Administration, Dosage Forms and Therapeutic Class APIs by drug code}
\usage{
use_drugCodes(
  drug_code,
  active_param = NULL,
  progress_bar = TRUE,
  alert_val_not_found = TRUE
)
}
\arguments{
\item{drug_code}{Drug code (numeric, string or vector)}

\item{active_param}{Only return dosage forms, schedules and routes of administration that are active. Default is NULL. Use 'yes' to returns parameter values that have a date that is greater than today or no date.}

\item{progress_bar}{Show progress bar. If TRUE, progress bar shows (default). Use FALSE to hide.}

\item{alert_val_not_found}{Show alert if a drug code is not in a component API. If TRUE, alerts show (default). Use FALSE to hide.}
}
\description{
Query a group of APIs that can only be queried by drug code and return a data frame.
}
\examples{
drug_codes_list <- c(0, 1, 8, 35413)
use_drugcodes(drug_codes_list)
}
