% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_api_response}
\alias{validate_api_response}
\title{Validate response of API call}
\usage{
validate_api_response(GET_response)
}
\arguments{
\item{GET_response}{Response object from \code{\link[httr:GET]{httr::GET()}}}
}
\description{
Prevents errors in API calls (i.e., querying for a value that doesn't exist, '404', etc....) from stopping the code from running.
}
\seealso{
\code{\link[httr:GET]{httr::GET()}}
}
