% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_products.R
\name{drugproduct_api}
\alias{drugproduct_api}
\title{Query Drug Product API of the Drug Product Database}
\usage{
drugproduct_api(
  drug_code = NULL,
  drug_identification_number = NULL,
  brand_name = NULL,
  status = NULL
)
}
\arguments{
\item{drug_code}{Code of the drug product.}

\item{drug_identification_number}{Drug Identification Number (DIN) of the drug product as a string.}

\item{brand_name}{Brand name of the drug product.}

\item{status}{Drug product status code. See \link{status_mapping} dataset for the description of each status code.}
}
\value{
Response from API call as a list. Information on DIN, brand name, company name, class name, therapeutic class name, Active Ingredient Group number and number of active ingredients, last date updated and additional details of drug product.
}
\description{
A function that queries Drug Product API using drug code, DIN, brand name, and  status.
Results can be unpredictable if drug code, DIN and brand name search criteria (arguments) are used in conjunction. Recommended to use only one.
The status argument can be used safely in combination with any of the other criteria
}
\examples{
drugproduct_api() # all drug products in the DPD
drugproduct_api(drug_code = 3847)
drugproduct_api(drug_identification_number = '00446564')
drugproduct_api(brand_name = 'apidra')
drugproduct_api(brand_name = 'apidra', status = 3) # Find Apidra products cancelled pre-market
}
\seealso{
\itemize{
\item \code{\link[=search_dp_api]{search_dp_api()}}, \code{\link[=query_dpd_by_din]{query_dpd_by_din()}}, \code{\link[=query_dpd_by_brandName]{query_dpd_by_brandName()}}
\item \link{status_mapping}
}
}
\author{
Ezinne MM Ndukwe
}
